ohos: Implement locales in qt speech engine
Populate available locales from the VoiceInfo.language field
returned by the TTS proxy's listVoices() method. Locales are
cached in QTextToSpeechEngineOhos::m_locales on engine
construction.

If available, the first locale from the list is used as the
initial m_currentLocale.

Filter availableVoices() by the current locale, as required by
QTextToSpeech::availableVoices(), which returns "the list of voices
available for the current locale". QTextToSpeech::findVoices() relies
on this: it iterates over the available locales and concatenates the
per-locale results.

Task-number: QTBUG-147179
Pick-to: 6.12
Change-Id: I98022805cead92e6d1764f619e300fc516f4ac90
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Zbigniew Chyla <zbigniew.chyla@qt.io>