WinRT: AudioSource - prevent duplicate scheduling of fetchMore
depending on the reading pattern of the QIODevice, fetchMore was
scheduled more than once, because it went through a timer.
This needs to be prevented to avoid messing up buffer logic. We
do not need the timer logic in this codepath, as fetchMore is
asynchronous, anyway.

Fixes: QTBUG-139660
Pick-to: 6.9 6.8
Change-Id: Ic1daa57304b81bacacf8b4337987b0f1ef10a047
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit e274f5b2b858b0d08b09bc30ce9bb02c93464d5f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>