Add security tags to the Qt TextToSpeech module
The platform/plugin-independent code, as well as those plugin
implementations that rely exclusively on system APIs to produce and play
the speech audio, do not contain any security critical code.

The two exceptions are the audio processors in the flite and the Windows
implementations. The former loads voice libraries using QLibrary, and
traverses the token tree. The latter inspects the PCM data to detect
segments of silence. This suggests that we would benefit from fuzz-
testing, sanitizer-enabled tests, and other scrutiny for this code, even
though the data we operate on generally comes from a trusted source.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-135774
Change-Id: Id6437cd2c8dc9c5d48a38c38ccc4a8c4275ef2f4
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>